make web server process auth requests
This adds the proper resolver functionality into the web server under
the `GET /request/:roleId/:action/:resource` route. It can respond
either in `text/plain` or `application/json`.

Currently some of the logic is a bit ugly as a lot of the `Either`s need
to be lifted into `ExceptT` with a single error datatype
(`ProgrammeError`) but for now it is functional.

Also remaining is to process actors rather than roles (i.e. to fetch all
the roles for a given actor).